// src/shared/api/types/gen/problemDetails.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * ViaTrade API
 * API для платформы инвестиционного анализа ViaTrade
 * OpenAPI spec version: v1
 */
import type { ProblemDetailsErrors } from './problemDetailsErrors';

export type ProblemDetails = {
	type: string;
	title: string;
	status: number;
	detail: string;
	instance: string;
	code: string;
	traceId: string;
	errors?: ProblemDetailsErrors;
};